decentralised_twitter/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
detoxify==0.5.1
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
filelock==3.9.0
huggingface-hub==0.12.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
numpy==1.24.1
packaging==23.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sentencepiece==0.1.97
sqlparse==0.4.3
tokenizers==0.12.1
torch==1.13.1
tqdm==4.64.1
transformers==4.22.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
